Modified resolve/completion for localrules/ruleorder: resolve works for all indexed elements, completion works for local files on 1st invocation, all files on further invocations.
