+++ src/main/resources/META-INF/plugin.xml
@@ -71,6 +71,10 @@
             language="Python"
             implementationClass="com.jetbrains.snakecharm.codeInsight.completion.SmkLambdaParameterInSectionCompletionContributor"
     />
+    <completion.contributor
+            language="Python"
+            implementationClass="com.jetbrains.snakecharm.codeInsight.completion.SmkRulesAndCheckpointsCompletionContributor"
+    />
 
     <localInspection
         language="Snakemake"
